+++ profiler/src/ini.c
@@ -176,7 +176,7 @@ static void split_data(ini_t *ini) {
 ini_t* ini_load(const char *filename) {
   ini_t *ini = NULL;
   FILE *fp = NULL;
-  int n, sz;
+  size_t n, sz;
 
   /* Init ini struct */
   ini = malloc(sizeof(*ini));
